pyproject.toml
[project]
name = "qwebsite-automation-bot"
version = "2.4.0"
description = "QwebsiteAutomationBot - Enterprise-grade Website Automation Agent with Business Intelligence"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Enigmatix", email = "enigmatix@example.com" }]
keywords = [
    "automation",
    "robocorp",
    "google-sheets",
    "business-intelligence",
    "api-integration",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Robocorp Framework
    "robocorp>=2.0.0",
    "robocorp-tasks>=3.0.0",
    "robocorp-workitems>=1.4.0",

    # Configuration & Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.8.1",

    # HTTP & API
    "requests>=2.31.0",

    # Environment
    "python-dotenv>=1.0.0",

    # Google API (JWT signing)
    "cryptography>=41.0.0",

    # Data Analysis & Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",

    # Advanced Analytics & Forecasting
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "responses>=0.24.0",
    "freezegun>=1.2.0",
    "ruff>=0.8.0",
    "pre-commit>=3.6.0",
]

[project.urls]
Homepage = "https://github.com/enigmatix/qwebsiteautomationbot"
Documentation = "https://github.com/enigmatix/qwebsiteautomationbot#readme"
Repository = "https://github.com/enigmatix/qwebsiteautomationbot.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "workflow"]

# ============================================
# Ruff Configuration (Linting & Formatting)
# ============================================
[tool.ruff]
line-length = 120
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
    "output",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "W",      # pycodestyle warnings
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "B904",   # raise from (too verbose for simple cases)
    "SIM105", # contextlib.suppress (readability preference)
    "SIM108", # ternary operator
    "SIM401", # dict.get instead of if block (readability preference)
    "RUF012", # mutable class attributes (common pattern)
    "UP038",  # isinstance tuple (Python 3.10+ union style optional)
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["src", "workflow", "processes", "libraries", "config"]
force-single-line = false
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

# ============================================
# Pytest Configuration
# ============================================
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "workflow", "."]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: marks tests as unit tests (fast, no external dependencies)",
    "integration: marks tests as integration tests (may require external services)",
    "slow: marks tests as slow running",
    "api: marks tests that interact with external APIs",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# ============================================
# Coverage Configuration
# ============================================
[tool.coverage.run]
source = ["src", "workflow"]
branch = true
omit = [
    "*/__pycache__/*",
    "*/tests/*",
    "*/.venv/*",
    "*/output/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "output/coverage_html"

# ============================================
# MyPy Configuration (Optional Type Checking)
# ============================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["tests", "output"]
